vyatta-cluster: wait for link up on clustering for up to 10 seconds
If the interfaces involved in a cluster are slow to either gain carrier
or show as link up during boot, the cluster configuration will fail to
load.  Add a delay of up to 10 seconds for this to occur, retesting
every second for a change in carrier / link status.

Bug #243 http://bugzilla.vyos.net/show_bug.cgi?id=243